+++ config/simple.js
@@ -1,0 +1,11 @@
+'use strict';
+
+module.exports = [
+    {
+        method: 'GET',
+        path: '/hello-motto',
+        handler: function (request, reply){
+            return "Hello Motorola";
+        }
+    },
+]
